+++ backend/src/nodes/impl/dithering/common.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 from ..image_utils import MAX_VALUES_BY_DTYPE
 
 
@@ -12,9 +13,7 @@ def float_to_dtype(image: np.ndarray, dtype: np.dtype) -> np.ndarray:
     return (image * max_value).astype(dtype)
 
 
-def apply_to_all_channels(
-    one_channel_filter, image: np.ndarray, *args, **kwargs
-) -> np.ndarray:
+def apply_to_all_channels(one_channel_filter, image: np.ndarray, *args, **kwargs) -> np.ndarray:
     if image.ndim == 2:
         return one_channel_filter(image, *args, **kwargs)
     output_image = np.stack(
