+++ backend/src/nodes/impl/dithering/palette.py
@@ -1,5 +1,6 @@
-import numpy as np
 import cv2
+import numpy as np
+from sanic.log import logger
 
 from .common import dtype_to_float
 from ..image_utils import as_3d
@@ -16,7 +17,7 @@ def distinct_colors(image: np.ndarray) -> np.ndarray:
 def kmeans_palette(image: np.ndarray, num_colors: int) -> np.ndarray:
     if image.ndim == 2:
         image = as_3d(image)
-    flat_image = dtype_to_float(image.reshape((-1,image.shape[2])))
+    flat_image = dtype_to_float(image.reshape((-1, image.shape[2])))
 
     max_iter = 10
     epsilon = 1.0
@@ -26,3 +27,39 @@ def kmeans_palette(image: np.ndarray, num_colors: int) -> np.ndarray:
     ret, label, center = cv2.kmeans(flat_image, num_colors, None, criteria, attempts, cv2.KMEANS_PP_CENTERS)
 
     return center.reshape((1, -1, image.shape[2]))
+
+
+class MedianCut:
+    def __init__(self, data: np.ndarray):
+        self.data = data
+        self.n_pixels, self.n_channels = data.shape
+        self.min_values = np.min(data, axis=0)
+        self.max_values = np.max(data, axis=0)
+        self.channel_ranges = self.max_values - self.min_values
+        self.biggest_range = np.max(self.channel_ranges)
+
+    def split(self):
+        widest_channel = np.argmax(self.channel_ranges)
+        median = np.median(self.data[:, widest_channel])
+        mask = self.data[:, widest_channel] > median
+        return MedianCut(self.data[mask == True]), MedianCut(self.data[mask == False])
+
+    def average(self):
+        return np.mean(self.data, axis=0)
+
+
+def median_cut(image: np.ndarray, num_colors: int) -> np.ndarray:
+    if image.ndim == 2:
+        image = as_3d(image)
+    flat_image = dtype_to_float(image.reshape((-1, image.shape[2])))
+
+    buckets = [MedianCut(flat_image)]
+    while len(buckets) < num_colors:
+        bucket_idx, bucket = max(enumerate(buckets), key=lambda x: x[1].biggest_range)
+        if bucket.n_pixels == 1:
+            break
+        buckets.pop(bucket_idx)
+
+        buckets.extend(bucket.split())
+
+    return np.stack([bucket.average() for bucket in buckets], axis=0).reshape((1,-1,image.shape[2]))
