+++ backend/src/nodes/impl/dithering/riemersma.py
@@ -5,8 +5,8 @@ import numpy as np
 
 from .color_distance import ColorDistanceFunction, nearest_palette_color
 from .common import apply_to_all_channels, dtype_to_float, float_to_dtype
-from .quantize import one_channel_nearest_uniform_color
 from .hilbert import HilbertCurve
+from .quantize import one_channel_nearest_uniform_color
 from ..image_utils import as_3d
 
 
@@ -21,14 +21,13 @@ def error_sum(history: deque, decay_ratio: float, value_type):
     b = math.e ** (math.log(decay_ratio) / (history.maxlen - 1))
     z = value_type()
     for i, x in enumerate(history):
-        z += x * b**i
+        z += x * b ** i
     return z
 
 
 def one_channel_riemersma_dither(
-    image: np.ndarray, history_length: int, decay_ratio: float, num_colors: int
+        image: np.ndarray, history_length: int, decay_ratio: float, num_colors: int
 ) -> np.ndarray:
-
     curve_size = next_power_of_two(max(image.shape))
 
     original_dtype = image.dtype
@@ -48,7 +47,7 @@ def one_channel_riemersma_dither(
 
 
 def riemersma_dither(
-    image: np.ndarray, history_length: int, decay_ratio: float, num_colors: int
+        image: np.ndarray, history_length: int, decay_ratio: float, num_colors: int
 ) -> np.ndarray:
     return apply_to_all_channels(
         one_channel_riemersma_dither,
@@ -60,11 +59,11 @@ def riemersma_dither(
 
 
 def nearest_color_riemersma_dither(
-    image: np.ndarray,
-    palette: np.ndarray,
-    color_distance_function: ColorDistanceFunction,
-    history_length: int,
-    decay_ratio: float,
+        image: np.ndarray,
+        palette: np.ndarray,
+        color_distance_function: ColorDistanceFunction,
+        history_length: int,
+        decay_ratio: float,
 ) -> np.ndarray:
     if image.ndim == 2:
         image = as_3d(image)
